Transferred script
I moved the script from the data folder to the main directory of the GitHub. I adapted a few lines to change to the proper directory so the script can properly access the data.

I changed the numbering of the scripts to include this percentile calculation before the main exploratory analyses.