fixed babylanguagelabdata
Changed filename and labid to match proper labid.
More importantly, it was discovered that for any participants with data at both 18 and 24 months, the raw and percentile scores were swapped between the 18 and 24 month old rows. This unfortunately had to be fixed by hand, but it has been carefully confirmed against the original webcdi dataset.