fixes
The if statement wasn't running because it was receiving several values. Fixed it so that if at least one value is found to be above 1, all values will be divided by 100. This is because in a beta regression the values of the dependent variable must be between 1 and 0.

Also removed an extra null model.

Made it so the session always starts in the working directory of the script file.